+++ test/sms_pmtct.test.js
@@ -1,5 +1,5 @@
 var vumigo = require('vumigo_v02');
-var fixtures = require('./fixtures');
+var fixtures = require('./fixtures_pmtct');
 var AppTester = vumigo.AppTester;
 var assert = require('assert');
